Basic Refactoring of Main, Model, And Window
